Model of a cheese counter
The problem solved using this Java file is provided below.

To restore order, the management installs a ticket machine that issues tickets to
customers. Tickets are numbered in the range 1..MT. When ticket MT has been issued,
the next ticket to be issued is ticket number 1, i.e. the management install a new ticket
roll. The cheese counter has a display that indicates the ticket number of the customer
currently being served. The customer with the ticket with the same number as the counter
display then goes to the counter and is served. When the service is finished, the number is
incremented (modulo MT). Model this system (with FSP) and show that, even when their
requests have low priority, meek customers are now served.